Exo 2 toujours en cours

+++ src/main/java/fr/univ_amu/iut/EmployeOrdinaire.java
@@ -3,7 +3,7 @@ package fr.univ_amu.iut;
 import java.time.LocalDate;
 
 public class EmployeOrdinaire extends Employe {
-    public EmployeOrdinaire(int numEmpl, String numSecu, String nom, String prenom, int echelon, LocalDate dateNaissance, LocalDate dateEmbauche, int tauxHorraire, int nbrHeure) {
+    public EmployeOrdinaire(int numEmpl, String numSecu, String nom, String prenom, int echelon, LocalDate dateNaissance, LocalDate dateEmbauche, double tauxHorraire, double nbrHeure) {
         super(numEmpl, numSecu, nom, prenom, echelon, dateNaissance, dateEmbauche, tauxHorraire, nbrHeure);
     }
 
@@ -14,4 +14,5 @@ public class EmployeOrdinaire extends Employe {
     private double getSalaireBrut(){
         return super.getSalaireBrute() + (super.getEchelon() * 100);
     }
+
 }
